move save button to the bottom to make space for project single close
